Add Svelte best practices documentation and implement workspace shell context
- Created documentation for function bindings, keyed each blocks, hydratable API, snippets, and Svelte reactivity.
- Implemented `WorkspaceShellContext` interface for managing workspace state and actions.
- Developed layout and page components for workspace, including ticket management and board creation.
- Added onboarding page for workspace selection with error handling.
- Introduced basic structure for board view with a placeholder message.
