+++ src/routes/(workspace)/board/+page.svelte
@@ -1,0 +1,13 @@
+<div class="flex h-full items-center justify-center p-8">
+    <div
+        class="w-full max-w-2xl rounded-xl border border-border/80 bg-surface-panel/80 p-6 text-center"
+    >
+        <h1 class="font-display text-lg font-semibold">
+            Board View Coming Soon
+        </h1>
+        <p class="mt-2 text-sm text-muted-foreground">
+            This route will host a focused per-board experience in a follow-up
+            step.
+        </p>
+    </div>
+</div>
