+++ src/routes/+layout.svelte
@@ -1,9 +1,49 @@
 <script lang="ts">
 	import "./layout.css";
+	import { goto } from "$app/navigation";
+	import { page } from "$app/stores";
 	import AppToolbar from "$lib/components/app/Toolbar/AppToolbar.svelte";
 	import { toolbarState } from "$lib/hooks/toolbar.svelte.js";
+	import { useWorkspace } from "$lib/hooks/workspace.svelte";
 
+	const workspace = useWorkspace();
 	const { children } = $props();
+	const currentPath = $derived.by(() => $page.url.pathname);
+
+	const bypassBootGate = $derived.by(
+		() =>
+			currentPath.startsWith("/onboarding") ||
+			currentPath.startsWith("/debug"),
+	);
+
+	function setBootFailure(context: string, error: unknown) {
+		const detail = error instanceof Error ? error.message : String(error);
+		console.error(`${context}: ${detail}`);
+	}
+
+	$effect(() => {
+		void workspace.init().catch((error) => {
+			setBootFailure("workspace.init failed", error);
+		});
+	});
+
+	$effect(() => {
+		if (
+			workspace.status === "ready" &&
+			currentPath.startsWith("/onboarding")
+		) {
+			void goto("/", { replaceState: true });
+			return;
+		}
+
+		if (
+			(workspace.status === "no_workspace" ||
+				workspace.status === "error") &&
+			!bypassBootGate
+		) {
+			void goto("/onboarding", { replaceState: true });
+		}
+	});
 
 	$effect(() => {
 		document.documentElement.classList.add("dark");
